refactor: decompose @blecsd/ai widget files into subdirectories
